ref: moved desktop sidebar components into /desktop dir
