+++ src/components/root/Navigation/SideBar.tsx
@@ -1,4 +1,4 @@
-import { DesktopSidebar } from '@/components/root/Navigation/DesktopSideBar'
+import { DesktopSidebar } from '@/components/root/Navigation/desktop/DesktopSideBar'
 import { ReactNode } from 'react'
 import { SidebarStoreProvider, SidebarStoreProviderProps } from '@/components/root/Navigation/SidebarStoreProvider'
 import MobileSideBar from '@/components/root/Navigation/mobile/MobileSideBar'
